backend/requirements.txt
flask==3.0.0
openai==1.3.7
firebase-admin==6.4.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
speechrecognition==3.10.0
pydub==0.25.1
requests==2.31.0
gunicorn==21.2.0
